Move TableDemoRobotAPI to src
